add evendata.s to align end of data segment avoiding relocate crash
if the data segments ends with an odd address, the loader failed on plain 68000 since the relocation table ended up on an odd address as well. On 68000, this causes a crash on driver relocation (since the first entry of the relocation table is supposed to be a long word). Not sure who's fault that is (GNU linker?).
This file ensures that the data segment end address is even.