Set font->distance from face->size->metrics ascender and descender
After calling `FT_Set_Char_Size()`` the face->size->metrics.ascender and
descender fields are filled in with what look like the top and bottom
distances, in pixels.

Using these gives very consistent baseline alignment across multiple
fonts.

The face->size->metrics.height field is effectively the line height, so is
not what is needed for the font->height field, which looks like it is the
height of the MFDB buffer instead. So we set this to the sum of the
ascender and descender.