attempt to fix "ragged" wide line circles
Circles are drawn as series of line segments. With wide lines, the line
segments are replaced by rectangles connected at the center points of
the smaller side. For small circle radii, this results in a "ragged"
appearance as the rectangle gape open at the circle circumference. To fix
that, we now draw small circles (with half the line width radius) at the
connection points